updated bpp_predicates_map type to u8

+++ include/fuzzmax.h
@@ -4,20 +4,22 @@
 
 #include "types.h"
 
-
 typedef struct fuzzmax_shmem {
-  s32 shm_id;
+  s32                   shm_id;
   struct fuzzmax_shmem *map;
-//  size_t map_size;
-  
+  //  size_t map_size;
+
   u32 fuzzmax_counter;
-  u8 histogram[256]; // TODO: update the max predicate allowed to large value later
-  u32 bpp_predicates_map[32896];
+  u8  histogram[256];  // TODO: update the max predicate allowed to large value
+                       // later
+  u8 bpp_predicates_map[32896];
 
 } fuzzmax_shmem_t;
 
-void update_fuzzmax_shmem(fuzzmax_shmem_t *__afl_fuzzmax_ptr, u32* T, u32 npreds);
-fuzzmax_shmem_t  *afl_fuzzmax_shm_init();
-void afl_fuzzmax_shm_deinit(fuzzmax_shmem_t *shm);
+void update_fuzzmax_shmem(fuzzmax_shmem_t *__afl_fuzzmax_ptr, u8 *T,
+                          u32 npreds);
+
+fuzzmax_shmem_t *afl_fuzzmax_shm_init();
+void             afl_fuzzmax_shm_deinit(fuzzmax_shmem_t *shm);
 
 #endif  // _AFL_FUZZMAX_H
